requirements.txt
-e .

# Django environment.
# django[argon2,bcrypt]  # See tools/install_django.sh
djangorestframework==3.14.*
dj-database-url
django-dotenv

# PostgreSQL testing.
psycopg2-binary

# Packaging.
twine
wheel

# Tooling.
black==23.9.1
django-test-migrations==1.3.0
mkdocs==1.5.2
mkdocs-material==9.3.1
pymdown-extensions==10.3
mypy==1.5.1
pytest==7.4.2
pytest-django==4.5.2
pytest-cov
ruff==0.0.289
